Deadband checking changed

+++ src/mappings/FloatMapping.h
@@ -40,7 +40,7 @@ public:
      * @param deadbandValue indicates a change in value of the register that is insignificant data.
      */
     FloatMapping(const std::string& reference, RegisterType registerType, const std::vector<int32_t>& addresses,
-                 bool readRestricted = false, int16_t slaveAddress = -1, float deadbandValue = 0.0);
+                 bool readRestricted = false, int16_t slaveAddress = -1, double deadbandValue = 0.0);
 
     /**
      * @details Override methods will be executed on devices reading thread, so that this parsing can be done
