+++ src/mappings/Int16Mapping.cpp
@@ -25,7 +25,7 @@
 namespace wolkabout
 {
 Int16Mapping::Int16Mapping(const std::string& reference, RegisterMapping::RegisterType registerType, int32_t address,
-                           bool readRestricted, int16_t slaveAddress, float deadbandValue)
+                           bool readRestricted, int16_t slaveAddress, double deadbandValue)
 : RegisterMapping(reference, registerType, address, OutputType::INT16, readRestricted, slaveAddress, deadbandValue)
 {
     // Stood here, but is actually redundant.
