+++ src/mappings/UInt16Mapping.cpp
@@ -23,7 +23,7 @@
 namespace wolkabout
 {
 UInt16Mapping::UInt16Mapping(const std::string& reference, RegisterMapping::RegisterType registerType, int32_t address,
-                             bool readRestricted, int16_t slaveAddress, float deadbandValue)
+                             bool readRestricted, int16_t slaveAddress, double deadbandValue)
 : RegisterMapping(reference, registerType, address, readRestricted, slaveAddress, deadbandValue)
 {
     if (!(registerType == RegisterType::INPUT_REGISTER || registerType == RegisterType::HOLDING_REGISTER))
