+++ src/mappings/UInt16Mapping.h
@@ -40,7 +40,7 @@ public:
      * @param deadbandValue indicates a change in value of the register that is insignificant data.
      */
     UInt16Mapping(const std::string& reference, RegisterType registerType, int32_t address, bool readRestricted = false,
-                  int16_t slaveAddress = -1, float deadbandValue = 0.0);
+                  int16_t slaveAddress = -1, double deadbandValue = 0.0);
 
     /**
      * @details Override methods will be executed on devices reading thread, so that this parsing can be done
