+++ src/mappings/UInt32Mapping.h
@@ -42,7 +42,7 @@ public:
      */
     UInt32Mapping(const std::string& reference, RegisterType registerType, const std::vector<int32_t>& addresses,
                   OperationType operation, bool readRestricted = false, int16_t slaveAddress = -1,
-                  float deadbandValue = 0.0);
+                  double deadbandValue = 0.0);
 
     /**
      * @details Override methods will be executed on devices reading thread, so that this parsing can be done
